moved import data out
