+++ src/import.js
@@ -1,0 +1,1698 @@
+// Import Data Module - File selection, pattern matching, and import workflow
+
+export function createImportModule({
+	invoke,
+	open,
+	isFileAlreadyImported,
+	refreshExistingFilePaths,
+	loadParticipantsView,
+	loadFiles,
+	navigateTo,
+	setLastImportView,
+}) {
+	// Internal state
+	let selectedFolder = null
+	let currentFiles = []
+	let currentPattern = ''
+	let fileParticipantIds = {}
+	let selectedFiles = new Set()
+	let sortField = 'path'
+	let sortDirection = 'asc'
+	let columnWidths = {
+		path: 1200,
+		filename: 180,
+		participant: 150,
+	}
+	let isImportInProgress = false
+	let reviewFileMetadata = {}
+	let selectedReviewFiles = new Set()
+	let reviewSortField = 'path'
+	let reviewSortDirection = 'asc'
+	let autoParticipantIds = {} // Auto-extracted IDs for the current pattern
+	let patternInputDebounce = null
+
+	function getFileExtensions() {
+		const select = document.getElementById('file-type-select')
+		const selected = Array.from(select.selectedOptions).map((opt) => opt.value)
+
+		// If custom is selected, get the custom extension
+		if (selected.includes('custom')) {
+			const customInput = document.getElementById('custom-ext-input')
+			let ext = customInput.value.trim()
+			if (ext.startsWith('*.')) {
+				ext = ext.substring(1)
+			}
+			// Remove 'custom' and add the actual extension
+			return selected.filter((v) => v !== 'custom').concat(ext ? [ext] : [])
+		}
+
+		return selected.filter((v) => v !== '') // Filter out empty option
+	}
+
+	function patternToRegex(pattern) {
+		if (!pattern) return null
+
+		// Handle special tokens that don't use regex
+		if (
+			pattern === '{parent}' ||
+			pattern === '{dirname}' ||
+			pattern === '{dir}' ||
+			pattern === '{filename}' ||
+			pattern === '{basename}' ||
+			pattern === '{id}/*'
+		) {
+			return null // These are handled separately
+		}
+
+		if (!pattern.includes('{id}')) return null
+
+		// Find what character comes after {id} to make matching non-greedy
+		const idIndex = pattern.indexOf('{id}')
+		const afterId = pattern.charAt(idIndex + 4) // Character after '{id}'
+
+		let characterClass
+		if (afterId === '_') {
+			// Exclude underscore from character class so it stops before delimiter
+			characterClass = '([a-zA-Z0-9\\-]+)'
+		} else if (afterId === '-') {
+			// Exclude hyphen from character class so it stops before delimiter
+			characterClass = '([a-zA-Z0-9_]+)'
+		} else if (afterId === '.') {
+			// Exclude period from character class so it stops before delimiter
+			characterClass = '([a-zA-Z0-9_\\-]+)'
+		} else {
+			// Default: capture typical ID characters greedily
+			characterClass = '([a-zA-Z0-9_\\-]+)'
+		}
+
+		let regex = pattern
+			.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
+			.replace(/\\\{id\\\}/g, characterClass)
+			.replace(/\\\*/g, '.*')
+
+		return new RegExp(regex)
+	}
+
+	function normalizeNamedGroupPattern(pattern) {
+		return pattern.replace(/\(\?P<([a-zA-Z0-9_]+)>/g, '(?<$1>')
+	}
+
+	function buildRegexFromString(pattern) {
+		if (!pattern) return null
+
+		try {
+			return new RegExp(normalizeNamedGroupPattern(pattern))
+		} catch (error) {
+			console.warn('Invalid regex pattern provided:', pattern, error)
+			return null
+		}
+	}
+
+	function extractIdFromPath(path, pattern) {
+		if (!pattern) return null
+
+		const normalized = pattern.trim()
+
+		// Handle special token patterns
+		if (
+			normalized === '{parent}' ||
+			normalized === '{dirname}' ||
+			normalized === '{dir}' ||
+			normalized === '{id}/*'
+		) {
+			const parts = path.split('/')
+			const parentDir = parts[parts.length - 2]
+			if (parentDir) {
+				const pathBeforeParent = parts.slice(0, -2).join('/') + '/'
+				return {
+					id: parentDir,
+					start: pathBeforeParent.length,
+					length: parentDir.length,
+					isDirectory: true,
+				}
+			}
+			return null
+		}
+
+		if (normalized === '{filename}') {
+			const filename = path.split('/').pop()
+			if (!filename) return null
+			const nameWithoutExt = filename.includes('.')
+				? filename.substring(0, filename.lastIndexOf('.'))
+				: filename
+			const dir = path.substring(0, path.lastIndexOf('/') + 1)
+			return {
+				id: nameWithoutExt,
+				start: dir.length,
+				length: nameWithoutExt.length,
+			}
+		}
+
+		if (normalized === '{basename}') {
+			const filename = path.split('/').pop()
+			if (!filename) return null
+			const dir = path.substring(0, path.lastIndexOf('/') + 1)
+			return {
+				id: filename,
+				start: dir.length,
+				length: filename.length,
+			}
+		}
+
+		if (normalized.startsWith('{parent:') && normalized.endsWith('}')) {
+			const parts = path.split('/')
+			if (parts.length < 2) return null
+			const parentDir = parts[parts.length - 2]
+			const innerPattern = normalized.slice('{parent:'.length, -1)
+			const pathBeforeParent = parts.slice(0, -2).join('/') + '/'
+
+			if (!parentDir) return null
+
+			if (innerPattern === '{id}') {
+				return {
+					id: parentDir,
+					start: pathBeforeParent.length,
+					length: parentDir.length,
+					isDirectory: true,
+				}
+			}
+
+			const parentRegex = patternToRegex(innerPattern)
+			if (!parentRegex) return null
+			const parentMatch = parentDir.match(parentRegex)
+			if (parentMatch && parentMatch[1]) {
+				const idStartInParent = parentMatch.index + parentMatch[0].indexOf(parentMatch[1])
+				return {
+					id: parentMatch[1],
+					start: pathBeforeParent.length + idStartInParent,
+					length: parentMatch[1].length,
+					isDirectory: true,
+				}
+			}
+			return null
+		}
+
+		if (normalized.startsWith('{stem:') && normalized.endsWith('}')) {
+			const filename = path.split('/').pop()
+			if (!filename) return null
+			const stem = filename.includes('.')
+				? filename.substring(0, filename.lastIndexOf('.'))
+				: filename
+			const dir = path.substring(0, path.lastIndexOf('/') + 1)
+			const innerPattern = normalized.slice('{stem:'.length, -1)
+
+			if (innerPattern === '{id}') {
+				return {
+					id: stem,
+					start: dir.length,
+					length: stem.length,
+				}
+			}
+
+			const stemRegex = patternToRegex(innerPattern)
+			if (!stemRegex) return null
+			const stemMatch = stem.match(stemRegex)
+			if (stemMatch && stemMatch[1]) {
+				const idStartInStem = stemMatch.index + stemMatch[0].indexOf(stemMatch[1])
+				return {
+					id: stemMatch[1],
+					start: dir.length + idStartInStem,
+					length: stemMatch[1].length,
+				}
+			}
+			return null
+		}
+
+		if (normalized.includes('{id}')) {
+			const filename = path.split('/').pop()
+			const regex = patternToRegex(normalized)
+
+			if (!regex || !filename) return null
+
+			const match = regex.exec(filename)
+			if (match && match[1]) {
+				const dir = path.substring(0, path.lastIndexOf('/') + 1)
+				const idStart = match.index + match[0].indexOf(match[1])
+				return { id: match[1], start: dir.length + idStart, length: match[1].length }
+			}
+			return null
+		}
+
+		const regex = buildRegexFromString(normalized)
+		if (!regex) return null
+		const match = regex.exec(path)
+		if (!match) return null
+
+		const groups = match.groups || {}
+		const extractedId = groups.id || match[1]
+		if (!extractedId) return null
+
+		const matchIndex = typeof match.index === 'number' ? match.index : path.indexOf(match[0])
+		const idStartWithinMatch = match[0].indexOf(extractedId)
+		const start = matchIndex + (idStartWithinMatch >= 0 ? idStartWithinMatch : 0)
+
+		return { id: extractedId, start, length: extractedId.length }
+	}
+
+	function highlightPath(path, pattern, fallbackId = '') {
+		const normalizedPattern = pattern ? pattern.trim() : ''
+		const extraction = normalizedPattern ? extractIdFromPath(path, normalizedPattern) : null
+
+		if (extraction && extraction.id) {
+			const before = path.substring(0, extraction.start)
+			const highlighted = path.substring(extraction.start, extraction.start + extraction.length)
+			const after = path.substring(extraction.start + extraction.length)
+			return `<span style="color: #666;">${before}</span><span class="highlight">${highlighted}</span><span style="color: #666;">${after}</span>`
+		}
+
+		const candidateId = fallbackId ? fallbackId.toString() : ''
+		if (candidateId) {
+			let index = path.indexOf(candidateId)
+			if (index === -1) {
+				index = path.toLowerCase().indexOf(candidateId.toLowerCase())
+			}
+			if (index !== -1) {
+				const before = path.substring(0, index)
+				const highlighted = path.substring(index, index + candidateId.length)
+				const after = path.substring(index + candidateId.length)
+				return `<span style="color: #666;">${before}</span><span class="highlight">${highlighted}</span><span style="color: #666;">${after}</span>`
+			}
+		}
+
+		const filename = path.split('/').pop()
+		const dir = path.substring(0, path.lastIndexOf('/') + 1)
+		return `<span style="color: #666;">${dir}</span>${filename}`
+	}
+
+	function getFileMetadata(filePath) {
+		const parts = filePath.split('/')
+		const fullFilename = parts[parts.length - 1]
+		const dir = parts.slice(0, -1).join('/') + '/'
+		const lastDotIndex = fullFilename.lastIndexOf('.')
+		const filename = lastDotIndex > 0 ? fullFilename.substring(0, lastDotIndex) : fullFilename
+		const extension = lastDotIndex > 0 ? fullFilename.substring(lastDotIndex) : ''
+
+		return { dir, filename, extension, fullFilename }
+	}
+
+	function sortFiles(files) {
+		const sorted = [...files]
+		sorted.sort((a, b) => {
+			let aVal, bVal
+
+			switch (sortField) {
+				case 'path':
+					aVal = a
+					bVal = b
+					break
+				case 'filename':
+					aVal = getFileMetadata(a).filename
+					bVal = getFileMetadata(b).filename
+					break
+				case 'extension':
+					aVal = getFileMetadata(a).extension
+					bVal = getFileMetadata(b).extension
+					break
+				case 'participant':
+					aVal = fileParticipantIds[a] || ''
+					bVal = fileParticipantIds[b] || ''
+					break
+				default:
+					return 0
+			}
+
+			const comparison = aVal.localeCompare(bVal)
+			return sortDirection === 'asc' ? comparison : -comparison
+		})
+
+		return sorted
+	}
+
+	function updateSortIndicators() {
+		const headers = document.querySelectorAll('.file-list-header div[data-sort]')
+		headers.forEach((header) => {
+			const indicator = header.querySelector('.sort-indicator')
+			if (header.dataset.sort === sortField) {
+				indicator.textContent = sortDirection === 'asc' ? '▲' : '▼'
+			} else {
+				indicator.textContent = ''
+			}
+		})
+	}
+
+	function setSortField(field) {
+		if (sortField === field) {
+			// Toggle direction
+			sortDirection = sortDirection === 'asc' ? 'desc' : 'asc'
+		} else {
+			// New field, default to ascending
+			sortField = field
+			sortDirection = 'asc'
+		}
+		renderFiles()
+	}
+
+	function updateColumnWidths() {
+		// Update header columns
+		const headerPath = document.querySelector('.file-list-header .col-path')
+		const headerFilename = document.querySelector('.file-list-header .col-filename')
+		const headerParticipant = document.querySelector('.file-list-header .col-participant')
+
+		if (headerPath) headerPath.style.width = `${columnWidths.path}px`
+		if (headerFilename) headerFilename.style.width = `${columnWidths.filename}px`
+		if (headerParticipant) headerParticipant.style.width = `${columnWidths.participant}px`
+
+		// Update all rows
+		document.querySelectorAll('.file-list li .col-path').forEach((el) => {
+			el.style.width = `${columnWidths.path}px`
+		})
+		document.querySelectorAll('.file-list li .col-filename').forEach((el) => {
+			el.style.width = `${columnWidths.filename}px`
+		})
+		document.querySelectorAll('.file-list li .col-participant').forEach((el) => {
+			el.style.width = `${columnWidths.participant}px`
+		})
+	}
+
+	function initColumnResizers() {
+		const resizers = document.querySelectorAll('.column-resizer')
+
+		resizers.forEach((resizer) => {
+			let startX, startWidth, column
+
+			resizer.addEventListener('mousedown', (e) => {
+				e.stopPropagation() // Prevent sort
+				column = resizer.dataset.col
+				startX = e.pageX
+				startWidth = columnWidths[column]
+
+				resizer.classList.add('resizing')
+				document.body.style.cursor = 'col-resize'
+				document.body.style.userSelect = 'none'
+
+				const onMouseMove = (e) => {
+					const diff = e.pageX - startX
+					const newWidth = Math.max(50, startWidth + diff)
+					columnWidths[column] = newWidth
+					updateColumnWidths()
+				}
+
+				const onMouseUp = () => {
+					resizer.classList.remove('resizing')
+					document.body.style.cursor = ''
+					document.body.style.userSelect = ''
+					document.removeEventListener('mousemove', onMouseMove)
+					document.removeEventListener('mouseup', onMouseUp)
+				}
+
+				document.addEventListener('mousemove', onMouseMove)
+				document.addEventListener('mouseup', onMouseUp)
+			})
+		})
+	}
+
+	function renderFiles() {
+		const fileList = document.getElementById('file-list')
+		fileList.innerHTML = ''
+
+		for (const file of Array.from(selectedFiles)) {
+			if (!currentFiles.includes(file) || isFileAlreadyImported(file)) {
+				selectedFiles.delete(file)
+			}
+		}
+
+		if (currentFiles.length === 0) {
+			const li = document.createElement('li')
+			li.textContent = 'No files found'
+			li.style.gridColumn = '1 / -1'
+			fileList.appendChild(li)
+			return
+		}
+
+		// Sort files
+		const sortedFiles = sortFiles(currentFiles)
+		const activePattern = currentPattern ? currentPattern.trim() : ''
+
+		sortedFiles.forEach((file) => {
+			const li = document.createElement('li')
+			const alreadyImported = isFileAlreadyImported(file)
+			const metadata = getFileMetadata(file)
+			const manualId = fileParticipantIds[file] || ''
+			const autoId = autoParticipantIds[file] || ''
+			const effectiveId = manualId || autoId || ''
+
+			// Checkbox column
+			const checkboxDiv = document.createElement('div')
+			checkboxDiv.className = 'col-checkbox'
+			const checkbox = document.createElement('input')
+			checkbox.type = 'checkbox'
+			checkbox.className = 'file-checkbox'
+			checkbox.checked = selectedFiles.has(file)
+			if (alreadyImported) {
+				checkbox.checked = false
+				checkbox.disabled = true
+				checkbox.title = 'File already imported'
+				selectedFiles.delete(file)
+				li.classList.add('already-imported')
+			}
+			checkbox.addEventListener('change', (e) => {
+				if (e.target.checked) {
+					selectedFiles.add(file)
+				} else {
+					selectedFiles.delete(file)
+				}
+				updateSelectAllCheckbox()
+				updateImportButton()
+				updateSelectedFileCount()
+			})
+			checkboxDiv.appendChild(checkbox)
+
+			// Path column (directory only) with highlighting
+			const pathDiv = document.createElement('div')
+			pathDiv.className = 'file-path col-path'
+			pathDiv.innerHTML = highlightPath(file, activePattern, effectiveId)
+			pathDiv.title = file // Full path on hover
+			pathDiv.style.width = `${columnWidths.path}px`
+			if (alreadyImported) {
+				const badge = document.createElement('span')
+				badge.className = 'imported-badge'
+				badge.textContent = 'Imported'
+				pathDiv.appendChild(badge)
+			}
+
+			// Filename column
+			const filenameDiv = document.createElement('div')
+			filenameDiv.className = 'col-filename'
+			filenameDiv.textContent = metadata.filename
+			filenameDiv.title = metadata.fullFilename
+			filenameDiv.style.width = `${columnWidths.filename}px`
+
+			// Extension column
+			const extensionDiv = document.createElement('div')
+			extensionDiv.className = 'col-extension'
+			extensionDiv.textContent = metadata.extension
+
+			// Participant ID input column
+			const participantDiv = document.createElement('div')
+			participantDiv.className = 'col-participant'
+			participantDiv.style.width = `${columnWidths.participant}px`
+			const input = document.createElement('input')
+			input.type = 'text'
+			input.className = 'participant-id-input'
+			input.placeholder = 'Enter ID'
+
+			// Extract ID if pattern exists
+			if (manualId) {
+				input.value = manualId
+				input.classList.add('manual')
+				input.classList.remove('extracted')
+			} else if (autoId) {
+				input.value = autoId
+				input.classList.add('extracted')
+				input.classList.remove('manual')
+			} else {
+				input.value = ''
+				input.classList.remove('manual')
+				input.classList.remove('extracted')
+			}
+
+			// Update map when user edits
+			input.addEventListener('input', (e) => {
+				const value = e.target.value.trim()
+				if (value) {
+					fileParticipantIds[file] = value
+					delete autoParticipantIds[file]
+					input.classList.remove('extracted')
+					input.classList.add('manual')
+				} else {
+					delete fileParticipantIds[file]
+					input.classList.remove('manual')
+					input.classList.remove('extracted')
+					if (currentPattern) {
+						void applyPattern(currentPattern)
+					}
+				}
+				updateImportButton()
+			})
+			participantDiv.appendChild(input)
+
+			// Show in folder button
+			const actionsDiv = document.createElement('div')
+			actionsDiv.className = 'col-actions'
+			const showBtn = document.createElement('button')
+			showBtn.className = 'show-in-folder-btn'
+			showBtn.textContent = '📁'
+			showBtn.title = 'Show in Finder'
+			showBtn.addEventListener('click', async () => {
+				try {
+					await invoke('show_in_folder', { filePath: file })
+				} catch (error) {
+					console.error('Failed to show file in folder:', error)
+				}
+			})
+			actionsDiv.appendChild(showBtn)
+
+			li.appendChild(checkboxDiv)
+			li.appendChild(pathDiv)
+			li.appendChild(filenameDiv)
+			li.appendChild(extensionDiv)
+			li.appendChild(participantDiv)
+			li.appendChild(actionsDiv)
+			fileList.appendChild(li)
+		})
+
+		// Update sort indicators
+		updateSortIndicators()
+
+		document.getElementById('file-count').textContent = currentFiles.length
+		updateSelectAllCheckbox()
+		updateSelectedFileCount()
+		updateImportButton()
+	}
+
+	function markActivePattern(pattern) {
+		const normalized = pattern ? pattern.trim() : ''
+		const patternCards = document.querySelectorAll('.pattern-card')
+		patternCards.forEach((card) => {
+			const macroValue = card.dataset.macro
+			const regexValue = card.dataset.regex
+			const macroChip = card.querySelector('[data-pattern-type="macro"]')
+			const regexChip = card.querySelector('[data-pattern-type="regex"]')
+
+			const macroActive = normalized && macroValue === normalized
+			const regexActive = normalized && regexValue === normalized
+
+			card.classList.toggle('active', macroActive || regexActive)
+			if (macroChip) {
+				macroChip.classList.toggle('active', macroActive)
+			}
+			if (regexChip) {
+				regexChip.classList.toggle('active', regexActive)
+			}
+		})
+	}
+
+	async function applyPattern(pattern) {
+		const normalized = pattern ? pattern.trim() : ''
+		const patternInput = document.getElementById('custom-pattern')
+		if (patternInput && patternInput.value.trim() !== normalized) {
+			patternInput.value = normalized
+		}
+		currentPattern = normalized
+		markActivePattern(normalized)
+
+		if (!normalized || currentFiles.length === 0) {
+			autoParticipantIds = {}
+			renderFiles()
+			updateImportButton()
+			return
+		}
+
+		try {
+			const results = await invoke('extract_ids_for_files', {
+				files: currentFiles,
+				pattern: normalized,
+			})
+
+			autoParticipantIds = {}
+			Object.entries(results || {}).forEach(([filePath, value]) => {
+				if (value !== null && value !== undefined && `${value}`.trim() !== '') {
+					autoParticipantIds[filePath] = `${value}`.trim()
+				}
+			})
+		} catch (error) {
+			console.error('Failed to extract IDs for pattern:', error)
+			autoParticipantIds = {}
+		}
+
+		renderFiles()
+		updateImportButton()
+	}
+
+	async function copyToClipboard(text) {
+		try {
+			await navigator.clipboard.writeText(text)
+		} catch (error) {
+			console.error('Clipboard error:', error)
+			throw error
+		}
+	}
+
+	async function updatePatternSuggestions() {
+		const container = document.getElementById('pattern-suggestions')
+		if (!container) return
+
+		if (currentFiles.length === 0) {
+			container.innerHTML =
+				'<div class="pattern-empty">Select a folder and file type to detect patterns.</div>'
+			markActivePattern('')
+			return
+		}
+
+		container.innerHTML = '<div class="pattern-loading">Analyzing filenames for patterns…</div>'
+
+		let suggestions = []
+		try {
+			suggestions = await invoke('suggest_patterns', { files: currentFiles })
+		} catch (error) {
+			console.error('Failed to fetch pattern suggestions:', error)
+			container.innerHTML =
+				'<div class="pattern-error">Unable to detect patterns. Try a different folder or check the console for details.</div>'
+			markActivePattern(currentPattern ? currentPattern.trim() : '')
+			return
+		}
+
+		if (!suggestions || suggestions.length === 0) {
+			container.innerHTML =
+				'<div class="pattern-empty">No ID patterns detected. Try selecting a different file type or adjust your dataset.</div>'
+			markActivePattern(currentPattern ? currentPattern.trim() : '')
+			return
+		}
+
+		container.innerHTML = ''
+
+		suggestions.forEach((sugg) => {
+			const macroValue = (sugg.pattern || '').trim()
+			const regexValue = (sugg.regex_pattern || '').trim()
+
+			const card = document.createElement('div')
+			card.className = 'pattern-card'
+			card.dataset.macro = macroValue
+			card.dataset.regex = regexValue
+
+			const title = document.createElement('div')
+			title.className = 'pattern-card-title'
+			title.textContent = sugg.description
+			card.appendChild(title)
+
+			const chipRow = document.createElement('div')
+			chipRow.className = 'pattern-card-chips'
+
+			const macroChip = document.createElement('button')
+			macroChip.className = 'pattern-chip pattern-chip--macro'
+			macroChip.dataset.patternType = 'macro'
+			macroChip.textContent = macroValue
+			macroChip.title = 'Use macro pattern'
+			macroChip.addEventListener('click', () => {
+				void applyPattern(macroValue)
+			})
+
+			chipRow.appendChild(macroChip)
+
+			if (regexValue) {
+				const regexChip = document.createElement('button')
+				regexChip.className = 'pattern-chip pattern-chip--regex'
+				regexChip.dataset.patternType = 'regex'
+				regexChip.textContent = regexValue
+				regexChip.title = 'Use regex pattern'
+				regexChip.addEventListener('click', () => {
+					void applyPattern(regexValue)
+				})
+				chipRow.appendChild(regexChip)
+			} else {
+				const regexLabel = document.createElement('span')
+				regexLabel.className = 'pattern-chip pattern-chip--regex disabled'
+				regexLabel.dataset.patternType = 'regex'
+				regexLabel.textContent = 'Regex unavailable'
+				chipRow.appendChild(regexLabel)
+			}
+			card.appendChild(chipRow)
+
+			if (sugg.example) {
+				const exampleRow = document.createElement('div')
+				exampleRow.className = 'pattern-card-example'
+				exampleRow.innerHTML = `<span>Example:</span> <span>${sugg.example}</span>`
+				card.appendChild(exampleRow)
+			}
+
+			if (Array.isArray(sugg.sample_extractions) && sugg.sample_extractions.length > 0) {
+				const sampleList = document.createElement('ul')
+				sampleList.className = 'pattern-card-samples'
+				sugg.sample_extractions.slice(0, 2).forEach((sample) => {
+					const samplePath = sample?.path ?? sample?.[0]
+					const sampleId = sample?.participant_id ?? sample?.[1]
+					if (!samplePath || !sampleId) return
+					const item = document.createElement('li')
+					item.innerHTML = `<span class="pattern-card-sample-path">${samplePath}</span><span class="pattern-card-sample-id">→ ${sampleId}</span>`
+					sampleList.appendChild(item)
+				})
+				card.appendChild(sampleList)
+			}
+
+			container.appendChild(card)
+		})
+
+		markActivePattern(currentPattern ? currentPattern.trim() : '')
+	}
+
+	async function searchFiles() {
+		if (!selectedFolder) return
+
+		const extensions = getFileExtensions()
+		if (extensions.length === 0) {
+			currentFiles = []
+			renderFiles()
+			markActivePattern('')
+			autoParticipantIds = {}
+			fileParticipantIds = {}
+			await updatePatternSuggestions()
+			return
+		}
+
+		currentFiles = await invoke('search_txt_files', { path: selectedFolder, extensions })
+		currentPattern = ''
+		autoParticipantIds = {}
+		fileParticipantIds = {}
+		const patternInput = document.getElementById('custom-pattern')
+		if (patternInput) {
+			patternInput.value = ''
+		}
+
+		renderFiles()
+		markActivePattern(currentPattern)
+		await updatePatternSuggestions()
+	}
+
+	async function updateFileTypeDropdown() {
+		if (!selectedFolder) return
+
+		const extensions = await invoke('get_extensions', { path: selectedFolder })
+		const select = document.getElementById('file-type-select')
+
+		select.innerHTML = ''
+
+		extensions.forEach((ext) => {
+			const option = document.createElement('option')
+			option.value = ext.extension
+			option.textContent = `${ext.extension} (${ext.count})`
+			select.appendChild(option)
+		})
+
+		const customOption = document.createElement('option')
+		customOption.value = 'custom'
+		customOption.textContent = 'Custom...'
+		select.appendChild(customOption)
+
+		// Auto-select the first extension by default
+		if (extensions.length > 0) {
+			select.options[0].selected = true
+		}
+	}
+
+	async function pickFolder() {
+		const selected = await open({
+			directory: true,
+			multiple: false,
+		})
+
+		if (selected) {
+			selectedFiles.clear()
+			updateSelectedFileCount()
+			await refreshExistingFilePaths()
+			selectedFolder = selected
+			document.getElementById('selected-path').textContent = selected
+			await updateFileTypeDropdown()
+			await searchFiles()
+		}
+	}
+
+	function updateSelectAllCheckbox() {
+		const selectAllCheckbox = document.getElementById('select-all-files')
+		const selectableFiles = currentFiles.filter((file) => !isFileAlreadyImported(file))
+		if (selectableFiles.length === 0) {
+			selectAllCheckbox.checked = false
+			return
+		}
+		selectAllCheckbox.checked = selectableFiles.every((file) => selectedFiles.has(file))
+	}
+
+	function updateSelectedFileCount() {
+		const el = document.getElementById('selected-count')
+		if (el) {
+			el.textContent = selectedFiles.size
+		}
+	}
+
+	function updateImportButton() {
+		const btn = document.getElementById('import-btn')
+
+		// Check if any files are selected and all selected files have participant IDs
+		const selectedFilesArray = Array.from(selectedFiles)
+		const hasSelection = selectedFilesArray.length > 0
+		const allSelectedHaveIds =
+			hasSelection &&
+			selectedFilesArray.every((file) => {
+				const manual = fileParticipantIds[file]
+				const auto = autoParticipantIds[file]
+				const value = manual || auto
+				return value !== undefined && value !== null && `${value}`.trim() !== ''
+			})
+
+		btn.disabled = !allSelectedHaveIds
+	}
+
+	function resetImportState() {
+		console.log('🔄 Resetting import workflow state')
+
+		// Clear tracking structures
+		isImportInProgress = false
+		selectedFiles.clear()
+		selectedReviewFiles = new Set()
+		reviewFileMetadata = {}
+		fileParticipantIds = {}
+		currentFiles = []
+		selectedFolder = null
+		currentPattern = ''
+		reviewSortField = 'path'
+		reviewSortDirection = 'asc'
+		autoParticipantIds = {}
+		setLastImportView('import')
+
+		// Reset step 1 UI elements
+		const selectedPathEl = document.getElementById('selected-path')
+		if (selectedPathEl) {
+			selectedPathEl.textContent = 'No folder selected'
+		}
+
+		const selectAllFiles = document.getElementById('select-all-files')
+		if (selectAllFiles) {
+			selectAllFiles.checked = false
+		}
+
+		const fileCountEl = document.getElementById('file-count')
+		if (fileCountEl) {
+			fileCountEl.textContent = '0'
+		}
+
+		updateSelectedFileCount()
+
+		const fileListEl = document.getElementById('file-list')
+		if (fileListEl) {
+			fileListEl.innerHTML = ''
+		}
+
+		const patternContainer = document.getElementById('pattern-suggestions')
+		if (patternContainer) {
+			patternContainer.innerHTML = ''
+		}
+
+		const customPatternInput = document.getElementById('custom-pattern')
+		if (customPatternInput) {
+			customPatternInput.value = ''
+		}
+
+		markActivePattern('')
+
+		const fileTypeSelect = document.getElementById('file-type-select')
+		if (fileTypeSelect) {
+			fileTypeSelect.innerHTML = '<option value="">Select file type(s)...</option>'
+			Array.from(fileTypeSelect.options).forEach((option) => (option.selected = false))
+		}
+
+		const customExtension = document.getElementById('custom-extension')
+		if (customExtension) {
+			customExtension.style.display = 'none'
+		}
+
+		const customExtInput = document.getElementById('custom-ext-input')
+		if (customExtInput) {
+			customExtInput.value = ''
+		}
+
+		// Reset review step UI elements
+		const reviewTable = document.getElementById('review-files-table')
+		if (reviewTable) {
+			reviewTable.innerHTML = ''
+		}
+
+		const reviewCountEl = document.getElementById('review-file-count')
+		if (reviewCountEl) {
+			reviewCountEl.textContent = '0'
+		}
+
+		const reviewSelectAll = document.getElementById('select-all-review')
+		if (reviewSelectAll) {
+			reviewSelectAll.checked = false
+			reviewSelectAll.indeterminate = false
+		}
+
+		const reviewImportBtn = document.getElementById('review-import-btn')
+		if (reviewImportBtn) {
+			reviewImportBtn.disabled = false
+			reviewImportBtn.innerHTML = 'Import Files →'
+		}
+
+		const detectBtn = document.getElementById('detect-types-btn')
+		if (detectBtn) {
+			detectBtn.disabled = false
+			detectBtn.innerHTML = '🔍 Detect File Types'
+		}
+
+		const analyzeBtn = document.getElementById('analyze-types-btn')
+		if (analyzeBtn) {
+			analyzeBtn.disabled = false
+			analyzeBtn.innerHTML = '🧬 Analyze Files'
+		}
+
+		const progressDiv = document.getElementById('detection-progress')
+		const progressBar = document.getElementById('progress-bar')
+		const progressText = document.getElementById('progress-text')
+		if (progressDiv && progressBar && progressText) {
+			progressDiv.style.display = 'none'
+			progressBar.style.width = '0%'
+			progressText.textContent = ''
+		}
+
+		updateSelectAllCheckbox()
+		updateImportButton()
+		updateReviewSelectAllCheckbox()
+
+		// Ensure the import step is the visible state if we were on the review view
+		const importView = document.getElementById('import-view')
+		const importReviewView = document.getElementById('import-review-view')
+		if (importReviewView && importReviewView.classList.contains('active')) {
+			importReviewView.classList.remove('active')
+			if (importView) {
+				importView.classList.add('active')
+			}
+
+			const importTab = document.querySelector('.tab[data-tab="import"]')
+			if (importTab) {
+				document.querySelectorAll('.tab').forEach((t) => t.classList.remove('active'))
+				importTab.classList.add('active')
+			}
+		}
+
+		console.log('✅ Import workflow reset complete')
+	}
+
+	function goToReviewStep() {
+		if (selectedFiles.size === 0) return
+
+		selectedFiles.forEach((file) => {
+			if (!fileParticipantIds[file] && autoParticipantIds[file]) {
+				fileParticipantIds[file] = autoParticipantIds[file]
+			}
+		})
+
+		// Build file-to-ID mapping
+		const filesToImport = Array.from(selectedFiles)
+
+		// Initialize metadata for each file
+		reviewFileMetadata = {}
+		selectedReviewFiles = new Set()
+
+		console.log(
+			'🔍 fileParticipantIds at review time:',
+			JSON.stringify(fileParticipantIds, null, 2),
+		)
+		console.log(`🔍 Number of files to import: ${filesToImport.length}`)
+		console.log(`🔍 Number of participant IDs in map: ${Object.keys(fileParticipantIds).length}`)
+
+		filesToImport.forEach((file) => {
+			const participantId = fileParticipantIds[file] || null
+			console.log(`🔍 File: ${file}`)
+			console.log(`   participantId from map: ${participantId}`)
+
+			reviewFileMetadata[file] = {
+				participant_id: participantId,
+				data_type: 'Unknown',
+				source: null,
+				grch_version: null,
+				row_count: null,
+				chromosome_count: null,
+				inferred_sex: null,
+			}
+			selectedReviewFiles.add(file) // Select all by default
+		})
+
+		console.log('🔍 reviewFileMetadata initialized:', reviewFileMetadata)
+
+		// Show review view
+		showReviewView()
+	}
+
+	function sortReviewFiles(filePaths) {
+		return filePaths.sort((a, b) => {
+			const metaA = reviewFileMetadata[a]
+			const metaB = reviewFileMetadata[b]
+
+			let valA, valB
+
+			switch (reviewSortField) {
+				case 'path':
+					valA = a.toLowerCase()
+					valB = b.toLowerCase()
+					break
+				case 'data_type':
+					valA = (metaA.data_type || '').toLowerCase()
+					valB = (metaB.data_type || '').toLowerCase()
+					break
+				case 'source':
+					valA = (metaA.source || '').toLowerCase()
+					valB = (metaB.source || '').toLowerCase()
+					break
+				case 'grch_version':
+					valA = (metaA.grch_version || '').toLowerCase()
+					valB = (metaB.grch_version || '').toLowerCase()
+					break
+				case 'row_count':
+					valA = metaA.row_count || 0
+					valB = metaB.row_count || 0
+					break
+				case 'chromosome_count':
+					valA = metaA.chromosome_count || 0
+					valB = metaB.chromosome_count || 0
+					break
+				case 'inferred_sex':
+					valA = (metaA.inferred_sex || '').toLowerCase()
+					valB = (metaB.inferred_sex || '').toLowerCase()
+					break
+				default:
+					valA = a.toLowerCase()
+					valB = b.toLowerCase()
+			}
+
+			if (valA < valB) return reviewSortDirection === 'asc' ? -1 : 1
+			if (valA > valB) return reviewSortDirection === 'asc' ? 1 : -1
+			return 0
+		})
+	}
+
+	function setReviewSortField(field) {
+		if (reviewSortField === field) {
+			reviewSortDirection = reviewSortDirection === 'asc' ? 'desc' : 'asc'
+		} else {
+			reviewSortField = field
+			reviewSortDirection = 'asc'
+		}
+
+		updateReviewSortIndicators()
+		showReviewView()
+	}
+
+	function updateReviewSortIndicators() {
+		document.querySelectorAll('#import-review-view .sortable-header').forEach((header) => {
+			const indicator = header.querySelector('.sort-indicator')
+			const field = header.dataset.sortField
+
+			if (field === reviewSortField) {
+				indicator.textContent = reviewSortDirection === 'asc' ? ' ▲' : ' ▼'
+			} else {
+				indicator.textContent = ''
+			}
+		})
+	}
+
+	function showReviewView() {
+		document.getElementById('review-file-count').textContent =
+			Object.keys(reviewFileMetadata).length
+
+		const tbody = document.getElementById('review-files-table')
+		tbody.innerHTML = ''
+
+		const sortedFiles = sortReviewFiles(Object.keys(reviewFileMetadata))
+
+		sortedFiles.forEach((filePath) => {
+			const metadata = reviewFileMetadata[filePath]
+			const row = document.createElement('tr')
+			row.style.borderBottom = '1px solid #eee'
+			row.dataset.filePath = filePath
+
+			// Checkbox cell
+			const checkboxCell = document.createElement('td')
+			checkboxCell.style.padding = '10px'
+			checkboxCell.style.textAlign = 'center'
+			const checkbox = document.createElement('input')
+			checkbox.type = 'checkbox'
+			checkbox.checked = selectedReviewFiles.has(filePath)
+			checkbox.addEventListener('change', (e) => {
+				if (e.target.checked) {
+					selectedReviewFiles.add(filePath)
+				} else {
+					selectedReviewFiles.delete(filePath)
+				}
+				updateReviewSelectAllCheckbox()
+			})
+			checkboxCell.appendChild(checkbox)
+			row.appendChild(checkboxCell)
+
+			// File path cell with folder icon
+			const pathCell = document.createElement('td')
+			pathCell.style.padding = '10px'
+			pathCell.style.fontSize = '13px'
+			pathCell.style.display = 'flex'
+			pathCell.style.alignItems = 'center'
+			pathCell.style.gap = '8px'
+
+			const pathText = document.createElement('span')
+			pathText.textContent = filePath
+			pathText.title = filePath
+			pathText.style.flex = '1'
+			pathText.style.overflow = 'hidden'
+			pathText.style.textOverflow = 'ellipsis'
+			pathText.style.whiteSpace = 'nowrap'
+
+			const folderBtn = document.createElement('button')
+			folderBtn.textContent = '📁'
+			folderBtn.style.padding = '4px 8px'
+			folderBtn.style.background = '#f0f0f0'
+			folderBtn.style.border = '1px solid #ddd'
+			folderBtn.style.borderRadius = '4px'
+			folderBtn.style.cursor = 'pointer'
+			folderBtn.style.flexShrink = '0'
+			folderBtn.style.zIndex = '10'
+			folderBtn.style.position = 'relative'
+			folderBtn.addEventListener('click', async (e) => {
+				e.preventDefault()
+				e.stopPropagation()
+				console.log('🔍 REVIEW Folder button clicked for:', filePath)
+				try {
+					console.log('🔍 Calling show_in_folder with filePath:', filePath)
+					await invoke('show_in_folder', { filePath: filePath })
+					console.log('✅ show_in_folder succeeded')
+				} catch (error) {
+					console.error('❌ show_in_folder error:', error)
+					alert(`Error opening folder: ${error}`)
+				}
+			})
+
+			pathCell.appendChild(pathText)
+			pathCell.appendChild(folderBtn)
+			row.appendChild(pathCell)
+
+			// Data type dropdown
+			const dataTypeCell = document.createElement('td')
+			dataTypeCell.style.padding = '10px'
+			const dataTypeSelect = document.createElement('select')
+			dataTypeSelect.style.width = '100%'
+			dataTypeSelect.style.padding = '6px'
+			dataTypeSelect.innerHTML = `
+			<option value="Unknown" ${metadata.data_type === 'Unknown' ? 'selected' : ''}>Unknown</option>
+			<option value="Genotype" ${metadata.data_type === 'Genotype' ? 'selected' : ''}>Genotype</option>
+		`
+			dataTypeSelect.addEventListener('change', (e) => {
+				reviewFileMetadata[filePath].data_type = e.target.value
+				updateRowVisibility(row, e.target.value)
+				applyReviewRowState(row, reviewFileMetadata[filePath])
+			})
+			dataTypeCell.appendChild(dataTypeSelect)
+			row.appendChild(dataTypeCell)
+
+			// Source dropdown
+			const sourceCell = document.createElement('td')
+			sourceCell.style.padding = '10px'
+			sourceCell.className = 'genotype-field'
+			const sourceSelect = document.createElement('select')
+			sourceSelect.style.width = '100%'
+			sourceSelect.style.padding = '6px'
+			sourceSelect.innerHTML = `
+			<option value="">-</option>
+			<option value="Unknown" ${metadata.source === 'Unknown' ? 'selected' : ''}>Unknown</option>
+			<option value="23andMe" ${metadata.source === '23andMe' ? 'selected' : ''}>23andMe</option>
+			<option value="AncestryDNA" ${
+				metadata.source === 'AncestryDNA' ? 'selected' : ''
+			}>AncestryDNA</option>
+			<option value="Genes for Good" ${
+				metadata.source === 'Genes for Good' ? 'selected' : ''
+			}>Genes for Good</option>
+			<option value="Dynamic DNA" ${
+				metadata.source === 'Dynamic DNA' ? 'selected' : ''
+			}>Dynamic DNA</option>
+		`
+			sourceSelect.addEventListener('change', (e) => {
+				reviewFileMetadata[filePath].source = e.target.value || null
+				applyReviewRowState(row, reviewFileMetadata[filePath])
+			})
+			sourceCell.appendChild(sourceSelect)
+			row.appendChild(sourceCell)
+
+			// GRCh version dropdown
+			const grchCell = document.createElement('td')
+			grchCell.style.padding = '10px'
+			grchCell.className = 'genotype-field'
+			const grchSelect = document.createElement('select')
+			grchSelect.style.width = '100%'
+			grchSelect.style.padding = '6px'
+			grchSelect.innerHTML = `
+			<option value="">-</option>
+			<option value="Unknown" ${metadata.grch_version === 'Unknown' ? 'selected' : ''}>Unknown</option>
+			<option value="36" ${metadata.grch_version === '36' ? 'selected' : ''}>36</option>
+			<option value="37" ${metadata.grch_version === '37' ? 'selected' : ''}>37</option>
+			<option value="38" ${metadata.grch_version === '38' ? 'selected' : ''}>38</option>
+		`
+			grchSelect.addEventListener('change', (e) => {
+				reviewFileMetadata[filePath].grch_version = e.target.value || null
+				applyReviewRowState(row, reviewFileMetadata[filePath])
+			})
+			grchCell.appendChild(grchSelect)
+			row.appendChild(grchCell)
+
+			// Row count cell (hidden - not used during import)
+			const rowCountCell = document.createElement('td')
+			rowCountCell.style.display = 'none'
+			rowCountCell.style.padding = '10px'
+			rowCountCell.style.fontSize = '13px'
+			rowCountCell.style.color = '#666'
+			rowCountCell.textContent = metadata.row_count ? metadata.row_count.toLocaleString() : '-'
+			row.appendChild(rowCountCell)
+
+			// Chromosome count cell (hidden - not used during import)
+			const chromCountCell = document.createElement('td')
+			chromCountCell.style.display = 'none'
+			chromCountCell.style.padding = '10px'
+			chromCountCell.style.fontSize = '13px'
+			chromCountCell.style.color = '#666'
+			chromCountCell.textContent = metadata.chromosome_count || '-'
+			row.appendChild(chromCountCell)
+
+			// Inferred sex cell (hidden - not used during import)
+			const sexCell = document.createElement('td')
+			sexCell.style.display = 'none'
+			sexCell.style.padding = '10px'
+			sexCell.style.fontSize = '13px'
+			sexCell.style.color = '#666'
+			sexCell.style.fontWeight = metadata.inferred_sex ? '600' : 'normal'
+			sexCell.textContent = metadata.inferred_sex || '-'
+			row.appendChild(sexCell)
+
+			applyReviewRowState(row, metadata)
+			updateRowVisibility(row, metadata.data_type)
+			tbody.appendChild(row)
+		})
+
+		navigateTo('import-review')
+		updateReviewSelectAllCheckbox()
+		updateReviewSortIndicators()
+	}
+
+	function updateReviewSelectAllCheckbox() {
+		const selectAllCheckbox = document.getElementById('select-all-review')
+		const totalFiles = Object.keys(reviewFileMetadata).length
+		const selectedCount = selectedReviewFiles.size
+
+		selectAllCheckbox.checked = selectedCount === totalFiles && totalFiles > 0
+		selectAllCheckbox.indeterminate = selectedCount > 0 && selectedCount < totalFiles
+	}
+
+	function isReviewMetadataComplete(metadata) {
+		if (!metadata) return false
+		const filledDataType = metadata.data_type && metadata.data_type !== 'Unknown'
+		const filledSource = metadata.source && metadata.source !== 'Unknown'
+		const filledGrch = metadata.grch_version && metadata.grch_version !== 'Unknown'
+		return filledDataType && filledSource && filledGrch
+	}
+
+	function applyReviewRowState(row, metadata) {
+		if (!row) return
+		if (isReviewMetadataComplete(metadata)) {
+			row.classList.add('review-row-complete')
+		} else {
+			row.classList.remove('review-row-complete')
+		}
+	}
+
+	function updateRowVisibility(row, dataType) {
+		const genotypeFields = row.querySelectorAll('.genotype-field')
+		genotypeFields.forEach((field) => {
+			field.style.display = dataType === 'Genotype' ? '' : 'none'
+		})
+	}
+
+	async function detectFileTypes() {
+		if (selectedReviewFiles.size === 0) {
+			alert('Please select files to detect')
+			return
+		}
+
+		const btn = document.getElementById('detect-types-btn')
+		const progressDiv = document.getElementById('detection-progress')
+		const progressBar = document.getElementById('progress-bar')
+		const progressText = document.getElementById('progress-text')
+
+		// Show progress UI IMMEDIATELY before any work
+		btn.disabled = true
+		btn.innerHTML = '<span class="spinner"></span>Detecting...'
+		progressDiv.style.display = 'flex'
+		progressBar.style.width = '0%'
+		progressText.textContent = 'Starting detection...'
+
+		// Force UI update before CLI calls
+		await new Promise((resolve) => setTimeout(resolve, 10))
+
+		try {
+			const selectedFilesArray = Array.from(selectedReviewFiles)
+			const batchSize = 1000 // Large batches since detection is fast
+			const totalFiles = selectedFilesArray.length
+			let processed = 0
+
+			// Process in batches to show progress
+			for (let i = 0; i < selectedFilesArray.length; i += batchSize) {
+				const batch = selectedFilesArray.slice(i, i + batchSize)
+
+				progressText.textContent = `Detecting file types... ${processed}/${totalFiles}`
+				progressBar.style.width = `${(processed / totalFiles) * 100}%`
+
+				const detections = await invoke('detect_file_types', { files: batch })
+
+				// Update all metadata and UI for this batch
+				Object.keys(detections).forEach((filePath) => {
+					const detection = detections[filePath]
+					console.log(`🔍 Detection result for ${filePath}:`, detection)
+					if (reviewFileMetadata[filePath] && detection) {
+						console.log(
+							`📝 BEFORE update - participant_id: ${reviewFileMetadata[filePath].participant_id}, data_type: ${reviewFileMetadata[filePath].data_type}, source: ${reviewFileMetadata[filePath].source}`,
+						)
+						reviewFileMetadata[filePath].data_type = detection.data_type
+						reviewFileMetadata[filePath].source = detection.source
+						reviewFileMetadata[filePath].grch_version = detection.grch_version
+						console.log(
+							`📝 AFTER update - participant_id: ${reviewFileMetadata[filePath].participant_id}, data_type: ${reviewFileMetadata[filePath].data_type}, source: ${reviewFileMetadata[filePath].source}`,
+						)
+						updateRowInPlace(filePath)
+					}
+				})
+
+				processed += batch.length
+			}
+
+			progressText.textContent = `Complete! Detected ${totalFiles} file types`
+			progressBar.style.width = '100%'
+
+			console.log(`✅ Detected ${totalFiles} file types`)
+		} catch (error) {
+			alert(`Error detecting file types: ${error}`)
+			console.error('Detection error:', error)
+		} finally {
+			btn.disabled = false
+			btn.innerHTML = '🔍 Detect File Types'
+
+			// Hide progress bar after a short delay
+			setTimeout(() => {
+				progressDiv.style.display = 'none'
+				progressBar.style.width = '0%'
+			}, 2000)
+		}
+	}
+
+	function updateRowInPlace(filePath) {
+		// Find row by iterating and comparing dataset
+		const rows = document.querySelectorAll('#review-files-table tr')
+		let targetRow = null
+		for (const row of rows) {
+			if (row.dataset.filePath === filePath) {
+				targetRow = row
+				break
+			}
+		}
+
+		if (!targetRow) return
+
+		const metadata = reviewFileMetadata[filePath]
+
+		// Update data type dropdown
+		const dataTypeSelect = targetRow.querySelector('td:nth-child(3) select')
+		if (dataTypeSelect) {
+			dataTypeSelect.value = metadata.data_type
+			updateRowVisibility(targetRow, metadata.data_type)
+		}
+
+		// Update source dropdown
+		const sourceSelect = targetRow.querySelector('td:nth-child(4) select')
+		if (sourceSelect && metadata.source) {
+			sourceSelect.value = metadata.source
+		}
+
+		// Update grch version dropdown
+		const grchSelect = targetRow.querySelector('td:nth-child(5) select')
+		if (grchSelect && metadata.grch_version) {
+			grchSelect.value = metadata.grch_version
+		}
+
+		// Update row count (column 6)
+		const rowCountCell = targetRow.querySelector('td:nth-child(6)')
+		if (rowCountCell) {
+			rowCountCell.textContent = metadata.row_count ? metadata.row_count.toLocaleString() : '-'
+		}
+
+		// Update chromosome count (column 7)
+		const chromCountCell = targetRow.querySelector('td:nth-child(7)')
+		if (chromCountCell) {
+			chromCountCell.textContent = metadata.chromosome_count || '-'
+		}
+
+		// Update inferred sex (column 8)
+		const sexCell = targetRow.querySelector('td:nth-child(8)')
+		if (sexCell) {
+			sexCell.textContent = metadata.inferred_sex || '-'
+			sexCell.style.fontWeight = metadata.inferred_sex ? '600' : 'normal'
+		}
+
+		applyReviewRowState(targetRow, metadata)
+	}
+
+	async function finalizeImport() {
+		const btn = document.getElementById('review-import-btn')
+		const progressDiv = document.getElementById('detection-progress')
+		const progressBar = document.getElementById('progress-bar')
+		const progressText = document.getElementById('progress-text')
+
+		// Mark import as in progress
+		isImportInProgress = true
+
+		// Show progress UI IMMEDIATELY before any work
+		btn.disabled = true
+		btn.innerHTML = '<span class="spinner"></span>Importing...'
+		progressDiv.style.display = 'flex'
+		progressBar.style.width = '0%'
+		progressText.textContent = 'Preparing import...'
+
+		// Force UI update before CLI calls
+		await new Promise((resolve) => setTimeout(resolve, 10))
+
+		try {
+			const filesToImport = Object.keys(reviewFileMetadata)
+			const totalFiles = filesToImport.length
+			let allResults = {
+				success: true,
+				imported: 0,
+				updated: 0,
+				conflicts: [],
+				errors: [],
+			}
+
+			// Build fileMetadata object for ALL files (single call - instant import)
+			const fileMetadata = {}
+			filesToImport.forEach((file) => {
+				const meta = reviewFileMetadata[file]
+				fileMetadata[file] = {
+					participant_id: meta.participant_id,
+					data_type: meta.data_type,
+					source: meta.source,
+					grch_version: meta.grch_version,
+				}
+				console.log(
+					`🔍 Preparing import - file: ${file}, participant_id: ${meta.participant_id}, data_type: ${meta.data_type}`,
+				)
+			})
+
+			console.log('📦 Full fileMetadata being sent to Tauri:', fileMetadata)
+
+			progressText.textContent = `Adding files to queue...`
+			progressBar.style.width = '50%'
+
+			// Fast import - add all files to queue instantly (no hashing)
+			const result = await invoke('import_files_pending', {
+				fileMetadata: fileMetadata,
+			})
+
+			progressText.textContent = `Complete! Added ${totalFiles} files to queue`
+			progressBar.style.width = '100%'
+
+			// Update results
+			if (result.success) {
+				allResults.imported = totalFiles
+			} else {
+				allResults.success = false
+				if (result.conflicts) allResults.conflicts.push(...result.conflicts)
+				if (result.errors) allResults.errors.push(...result.errors)
+			}
+
+			if (allResults.success) {
+				// Metadata was saved during import via CSV
+				await loadParticipantsView()
+				await loadFiles()
+
+				// Navigate to files tab after successful import
+				setTimeout(() => {
+					resetImportState()
+					isImportInProgress = false
+					progressDiv.style.display = 'none'
+					progressBar.style.width = '0%'
+					navigateTo('files')
+				}, 1000)
+			} else {
+				const updateConflicts = confirm(
+					`Some files had conflicts.\nDo you want to update the files with conflicts?`,
+				)
+
+				if (updateConflicts) {
+					alert('Update functionality coming soon')
+				}
+
+				// Re-enable button on failure
+				isImportInProgress = false
+				btn.disabled = false
+				btn.innerHTML = 'Import Files →'
+				setTimeout(() => {
+					progressDiv.style.display = 'none'
+					progressBar.style.width = '0%'
+				}, 2000)
+			}
+		} catch (error) {
+			alert(`Error: ${error}`)
+			console.error('Import error:', error)
+
+			// Re-enable button on error
+			isImportInProgress = false
+			btn.disabled = false
+			btn.innerHTML = 'Import Files →'
+			setTimeout(() => {
+				progressDiv.style.display = 'none'
+				progressBar.style.width = '0%'
+			}, 2000)
+		}
+	}
+
+	function handleSelectAllFiles(checked) {
+		if (checked) {
+			currentFiles.forEach((file) => {
+				if (!isFileAlreadyImported(file)) {
+					selectedFiles.add(file)
+				}
+			})
+		} else {
+			selectedFiles.clear()
+		}
+		renderFiles()
+	}
+
+	function handleSelectAllReview(checked) {
+		const allFiles = Object.keys(reviewFileMetadata)
+		if (checked) {
+			allFiles.forEach((file) => selectedReviewFiles.add(file))
+		} else {
+			selectedReviewFiles.clear()
+		}
+		showReviewView()
+	}
+
+	function handleBulkDataTypeChange(value) {
+		if (!value) return
+		if (selectedReviewFiles.size === 0) {
+			alert('Please select files to update')
+			return
+		}
+		selectedReviewFiles.forEach((file) => {
+			reviewFileMetadata[file].data_type = value
+		})
+		showReviewView()
+	}
+
+	function handleBulkSourceChange(value) {
+		if (!value) return
+		if (selectedReviewFiles.size === 0) {
+			alert('Please select files to update')
+			return
+		}
+		selectedReviewFiles.forEach((file) => {
+			reviewFileMetadata[file].source = value
+		})
+		showReviewView()
+	}
+
+	function handleBulkGrchVersionChange(value) {
+		if (!value) return
+		if (selectedReviewFiles.size === 0) {
+			alert('Please select files to update')
+			return
+		}
+		selectedReviewFiles.forEach((file) => {
+			reviewFileMetadata[file].grch_version = value
+		})
+		showReviewView()
+	}
+
+	function handleCustomPatternInput(value) {
+		if (patternInputDebounce) {
+			clearTimeout(patternInputDebounce)
+		}
+		patternInputDebounce = setTimeout(() => {
+			patternInputDebounce = null
+			applyPattern(value)
+		}, 300)
+	}
+
+	function handleCustomPatternBlur(value) {
+		if (patternInputDebounce) {
+			clearTimeout(patternInputDebounce)
+			patternInputDebounce = null
+		}
+		applyPattern(value)
+	}
+
+	function handleCustomPatternKeydown(key, value) {
+		if (key === 'Enter') {
+			if (patternInputDebounce) {
+				clearTimeout(patternInputDebounce)
+				patternInputDebounce = null
+			}
+			applyPattern(value)
+		}
+	}
+
+	async function handleCopyPattern(pattern) {
+		try {
+			await copyToClipboard(pattern)
+		} catch (error) {
+			console.error('Failed to copy pattern:', error)
+		}
+	}
+
+	return {
+		pickFolder,
+		searchFiles,
+		renderFiles,
+		updateSelectAllCheckbox,
+		updateSelectedFileCount,
+		updateImportButton,
+		resetImportState,
+		goToReviewStep,
+		showReviewView,
+		detectFileTypes,
+		finalizeImport,
+		setSortField,
+		initColumnResizers,
+		setReviewSortField,
+		applyPattern,
+		updatePatternSuggestions,
+		markActivePattern,
+		updateFileTypeDropdown,
+		getFileExtensions,
+		handleSelectAllFiles,
+		handleSelectAllReview,
+		handleBulkDataTypeChange,
+		handleBulkSourceChange,
+		handleBulkGrchVersionChange,
+		handleCustomPatternInput,
+		handleCustomPatternBlur,
+		handleCustomPatternKeydown,
+		handleCopyPattern,
+		copyToClipboard,
+		getIsImportInProgress: () => isImportInProgress,
+		getSelectedFolder: () => selectedFolder,
+		getSelectedFiles: () => selectedFiles,
+		getReviewFileMetadata: () => reviewFileMetadata,
+		getSelectedReviewFiles: () => selectedReviewFiles,
+	}
+}
